dom/base/test/test_meta_viewport_multiple_tags.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>multilple meta viewport tags</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <meta name="viewport" content="width=980">
  <meta name="viewport" content="initial-scale=1,maximum-scale=1">
  <script src="viewport_helpers.js"></script>
</head>
<body>
  <script type="application/javascript">
    "use strict";

    add_task(async function multiple_viewport_tags() {
      await SpecialPowers.pushPrefEnv(scaleRatio(1.0));

      let info = getViewportInfo(800, 480);
      info = getViewportInfo(800, 480);
      is(info.defaultZoom, 1, "initial zoom is 1");
      is(info.width, 800, "width should be 800");
      is(info.height, 480, "height should be 480");
      is(info.maxZoom, 1, "maximum scale is 1");
    });
  </script>
</body>
</html>
